Final commit

+++ app/Http/Controllers/ListingController.php
@@ -5,37 +5,29 @@ namespace App\Http\Controllers;
 use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
-use MongoDB\Driver\Session;
 
 class ListingController extends Controller
 {
-
-    // show all listings
-    public function index(Request $request)
-    {
-
+    // Show all listings
+    public function index() {
         return view('listings.index', [
-            'heading' => 'lastes lisenings',
-            'listings' => Listing::latest()->filter(\request(['tag', 'search']))->paginate(6),
+            'listings' => Listing::latest()->filter(request(['tag', 'search']))->paginate(6)
         ]);
     }
 
-    // show create form
-
-    public function create()
-    {
-        return view('listings.create');
-    }
-
-    // show single listing
-    public function show(Listing $listing)
-    {
+    //Show single listing
+    public function show(Listing $listing) {
         return view('listings.show', [
             'listing' => $listing
         ]);
     }
 
-    //store listing data
+    // Show Create Form
+    public function create() {
+        return view('listings.create');
+    }
+
+    // Store Listing Data
     public function store(Request $request) {
         $formFields = $request->validate([
             'title' => 'required',
@@ -51,24 +43,28 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
 
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully!');
     }
 
-    // Show edit form
-
-    public function edit(Listing $listing)
-    {
+    // Show Edit Form
+    public function edit(Listing $listing) {
         return view('listings.edit', ['listing' => $listing]);
     }
 
-    //update listing data
+    // Update Listing Data
     public function update(Request $request, Listing $listing) {
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
-            'company' => ['required', Rule::unique('listings', 'company')],
+            'company' => ['required'],
             'location' => 'required',
             'website' => 'required',
             'email' => ['required', 'email'],
@@ -80,15 +76,24 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
-
         $listing->update($formFields);
 
         return back()->with('message', 'Listing updated successfully!');
     }
 
-    // Delete listing
-    public function destory(Request $request, Listing $listing) {
+    // Delete Listing
+    public function destroy(Listing $listing) {
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
-        return redirect('/')->with('message', 'Listing deleted successfullt');
+        return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    // Manage Listings
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
